requirements-cpu.txt
torch>=2.0.1
torchaudio>=2.0.2
torchvision>=0.15.2
openai-whisper>=20230314
pathlib